requirements.txt
pylint
pyspark
pandas
streamlit
kaggle
psycopg2-binary